% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Polygon_squared_regions.R
\name{build_Polygon_squared_regions}
\alias{build_Polygon_squared_regions}
\title{Build a \code{SpatialPolygon} object made of squared regions}
\usage{
build_Polygon_squared_regions(n_rows = 5, n_cols = 4)
}
\arguments{
\item{n_rows}{the number of rows of the map}

\item{n_cols}{the number of columns of the map}
}
\value{
A \code{SpatialPolygon} object.
}
\description{
This function creates a rectangular \code{SpatialPolygons} map divided into square regions. The user specifies the number of rows and columns to define the grid's layout.
}
\examples{
build_Polygon_squared_regions(n_rows = 5, n_cols = 4)

}
