% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Z.R
\name{convert.Z.matrix_to_vector}
\alias{convert.Z.matrix_to_vector}
\title{Z from matrix to vector
This function convert the selection matrix \code{Z} used in the model into a vector of length \code{n} stating to which each observation belong.}
\usage{
convert.Z.matrix_to_vector(Z)
}
\arguments{
\item{Z}{}
}
\description{
Z from matrix to vector
This function convert the selection matrix \code{Z} used in the model into a vector of length \code{n} stating to which each observation belong.
}
