% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Z.R
\name{convert.Z.vector_to_matrix}
\alias{convert.Z.vector_to_matrix}
\title{Z from vector to matrix}
\usage{
convert.Z.vector_to_matrix(z)
}
\arguments{
\item{z}{}
}
\description{
This function convert a vector of length \code{n} stating to which each observation belong into the selection matrix \code{Z} used in the model
}
