% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_label_switching.R
\name{display.label.switching}
\alias{display.label.switching}
\title{Display label switching correction}
\usage{
display.label.switching(values, colors = NULL, names.diseases = NULL)
}
\arguments{
\item{values}{the output of the function \code{remove.label.switching}}

\item{colors}{(optional) colors of the clusters in the traceplots}

\item{names.diseases}{(optional) the disease names}
}
\value{
a \code{ggplot} object
}
\description{
This function displays the relabelling of the cluster centroids $\mu_{k}$ based on the applied algorithms.
}
