% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Inverse transformation of the logit function}
\usage{
expit(x)
}
\arguments{
\item{p}{a vector of probability values.}
}
\value{
the value \code{exp(p)/(1+exp(p))}.
}
\description{
This function implements the inverse transformation of the logit function.
}
