% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate.simulations}
\alias{generate.simulations}
\title{Simulate maps using DAGAR}
\usage{
generate.simulations(
  spatial.map,
  K = NULL,
  d = NULL,
  X = NULL,
  Beta = NULL,
  Sigma = NULL,
  z = NULL,
  Mu = NULL,
  range.mu,
  range.Sigma,
  range.beta = 1,
  rho,
  prob.null.centroid = 0.5,
  scale.factor.variance = 0.05,
  correct.mean.prob = F,
  plot.map = F
)
}
\arguments{
\item{spatial.map}{A \code{SpatialPolygonsDataFrame} object with the map you want
to simulate from.}

\item{K}{Number of clusters (if \code{z} is passed, it is ignored).}

\item{d}{Number of variables (if \code{Sigma} is passed, it is ignored).}

\item{X}{(optional) a data.frame containing some possible covariates. If passed, each value of the matrix of regression coefficients will be generated using \code{rnorm(1, 0, range.beta)}.}

\item{Beta}{(optional) a matrix of regression coefficients. It must have \code{p} rows and \code{d} columns. If passed, the regression coefficients are no longer generated randomly.}

\item{Sigma}{(optional) correlation matrix of the variables.}

\item{z}{(optional) a vector of clustering labels. If passed, the clusters will not be generated randomly using the multinomial distribution, and \code{K} will be set equal to the number of unique values of \code{z}.}

\item{Mu}{(optional) a matrix of cluster centroids. It must have \code{K} rows and \code{d} columns. If passed, the cluster centroids are no longer generated randomly.}

\item{range.mu}{Range of the uniform distribution used in data generation.}

\item{range.Sigma}{Range of the uniform distribution used to generate the
correlation matrix Sigma.}

\item{range.beta}{(optional) the standard deviation of the Gaussian distribution used to generate the regression coefficients}

\item{rho}{\code{rho} parameter of the DAGAR.}

\item{prob.null.centroid}{Probability that the \code{K} centroids related to a certain variable are zero. This is used to simulate scenarios where not all the \code{d} variables are responsible of the formation of the \code{K} clusters.}

\item{scale.factor.variance}{Scaling factor for variance-covariance matrix
(default 0.05).}

\item{correct.mean.prob}{Use the stick-breaking formulation of the
multinomial distribution while simulating (default \code{FALSE}).}

\item{plot.map}{Plot the map during the simulation (default \code{FALSE}).}
}
\value{
The function returns the map and the simulated data.
}
\description{
Simulate maps using DAGAR.
}
\examples{
data(west_states_data)

sim1 <- generate.simulations(spatial.map = west_states_data,
K = 3,
d = 10,
Sigma = NULL,
range.mu = 0.5,
range.Sigma = 0.1,
rho = c(0.8, 0.9),
prob.null.centroid = 0.5,
scale.factor.variance = 0.05,
correct.mean.prob = F,
plot.map = T)


sim2 <- generate.simulations(spatial.map = west_states_data,
K = 4,
d = 3,
Sigma = NULL,
range.mu = 1,
rho = c(0.01, 0.455, 0.9),
range.Sigma = 0.2,
prob.null.centroid = 0.5,
scale.factor.variance = 0.07,
correct.mean.prob = F,
plot.map = T)

}
