% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_correlations.R
\name{marginal_correlations}
\alias{marginal_correlations}
\title{Marginal correlation of the outcome variables}
\usage{
marginal_correlations(values)
}
\arguments{
\item{values}{an object of class \code{perla}.}
}
\value{
an object of class \code{perla} that contains also \code{Corr}, an array of the same dimension of \code{Sigma} collecting the marginal correlation matrix at every MCMC iteration.
}
\description{
This function derives the posterior distributions of the marginal correlations across outcome variables based on the MCMC runs.
}
