% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perla.R
\name{perla}
\alias{perla}
\title{PEnalized Regression with Localities Aggregation (PERLA)}
\usage{
perla(
  y,
  W = NULL,
  K,
  R = 10^4,
  prior.rho = "const",
  p.spike = 0.5,
  rho.value = 0.99,
  mean.penalty = c(),
  include.Sigma = T,
  mu0 = NULL,
  Sigma0 = NULL,
  burnin = NULL,
  initialization = NULL
)
}
\arguments{
\item{y}{the input dataset. It can be either of class \code{matrix} or \code{SpatialPolygonsDataFrame}}

\item{K}{the number of clusters}

\item{R}{the number of MCMC iterations (default \code{10^4})}

\item{prior.rho}{set the type of prior on the rho parameter of the CAR. If \code{const}, it assumes a fixed value, that is \code{rho.value}.
If \code{disc}, it assumes that rho can be equal to \code{0} (with probability \code{p.spike}) or \code{rho.value}.
If \code{cont}, it assumes that rho is generated from a mixture of a \code{dbeta(x,2,18)} (with probability \code{p.spike}) or a \code{dbeta(x,18,2)}.}

\item{rho.value}{default \code{0.99}}

\item{mean.penalty}{If \code{c()}, only the global shrinkage is considered. If \code{mean.penalty} contains \code{"c"}, then cluster penalties are added. If \code{mean.penalty} contains \code{"d"}, then disease penalties are added. If \code{mean.penalty} contains \code{"cd"}, then both cluster and disease penalties are added.
If \code{NULL}, no penalization is considered and the prior of cluster mean vectors are d-variate Gaussian distributions, zero-centered and with covariance matrix \code{Sigma0}}

\item{include.Sigma}{if \code{T}, the \code{Sigma} matrix becomes part of the prior variance of the cluster centroids.}

\item{burnin}{a vector of indexes denoting the MCMC draws to be discarded. If \code{NULL}, then only the starting values are discarded and the algorithm will perform \code{R+1} iterations.}

\item{initialization}{it can be either an object of class \code{perla} or a list with starting point. In the last case, each list element must be named after the parameter to be initialized. Names include \code{Mu}, \code{Prob}, \code{Z}, \code{Sigma}, \code{Rho})}
}
\description{
This function estimates the PERLA model using the Markov Chain Monte Carlo algorithm.
}
