% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{recover.loglikelihood}
\alias{recover.loglikelihood}
\title{A posteriori log-likelihood computation}
\usage{
recover.loglikelihood(value, burnin = NULL)
}
\arguments{
\item{burnin}{}
}
\description{
This function returns the log-likelihood value at each iteration of the MCMC.
}
