% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_label_switching.R
\name{remove.label.switching}
\alias{remove.label.switching}
\title{Solve label switching}
\usage{
remove.label.switching(
  values,
  methods = c("PRA", "ECR"),
  loglikelihood.values = NULL,
  burnin = NULL
)
}
\arguments{
\item{values}{an object of class \code{perla}}

\item{methods}{default \verb{("PRA","ECR")}}

\item{loglikelihood.values}{the output of the function \code{loglikelihood.values}. If \code{NULL} and the methods \code{PRA} and/or \code{ECR} are evoced to solve the label switching phenomenon, then the function will automatically compute the loglikelihood values.}

\item{burnin}{a vector of indexes denoting the MCMC draws to be discarded (default \code{NULL} means every draw is kept)}
}
\value{
a list of two elements.
\itemize{
\item \code{results} contains the permuted clustering centroids and clustering probabilities for each method used to resolve the label switching;
\item \code{relabelling} is the output of the \code{label.switching} function.
}
}
\description{
This function can be applied to remove the label switching effect.
}
