% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_label_switching.R
\name{solve.label.switching}
\alias{solve.label.switching}
\title{Solve label switching}
\usage{
\method{solve}{label.switching}(
  values,
  methods = c("PRA", "ECR", "STEPHENS"),
  loglikelihood.values = NULL,
  burnin = NULL
)
}
\arguments{
\item{values}{an object of class \code{perla}}

\item{methods}{}

\item{burnin}{a vector of indexes denoting the MCMC draws to be discarded (default \code{NULL} means every draw is kept)}
}
\description{
This function can be applied to remove the label switching effect.
}
