% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rho_continuous_S&S.R
\name{update_rho_continuous}
\alias{update_rho_continuous}
\title{Metropolis-Hastrings update of the CAR parameter}
\usage{
update_rho_continuous(
  rho.old,
  psi,
  tau = 1,
  W,
  lambda_spike = 3,
  lambda_slab = 7,
  a_spike = 0.5,
  b_spike = 18,
  a_slab = 18,
  b_slab = 2,
  sigma.rho = 1
)
}
\arguments{
\item{sigma.rho}{}
}
\description{
This function receives a value of rho and generates a new value from its posterior distribution given by the spike & slab prior.
It uses a random walk Metropolis-Hastrings move.
}
