% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rho_discrete.R
\name{update_rho_discrete}
\alias{update_rho_discrete}
\title{Discrete rho update}
\usage{
update_rho_discrete(
  psi,
  tau = 1,
  W,
  lambda.spike = 1,
  lambda.slab = 1,
  rho.value = 0.99,
  detOmega = NULL
)
}
\arguments{
\item{psi}{the \verb{n x (K-1)} matrix of the \code{psi} values}

\item{rho.value}{(default \code{0.99}) it contains the value of rho different from 0}

\item{detOmega}{}
}
\description{
Discrete rho update
}
