% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tau.R
\name{update_tau}
\alias{update_tau}
\title{Posterior update of the CAR variance parameters via conjugate priors}
\usage{
update_tau(Psi, D, W, Rho, shape, scale)
}
\arguments{
\item{Psi}{a matrix of dimension \verb{n x (K-1)}, where \code{n} is the number of spatial areas in the analysed map, and \code{K} is the number of clusters assumed.}

\item{D}{the diagonal matrix containing the number of neighbours.}

\item{W}{the proximity matrix.}

\item{Rho}{a vector of length \code{K-1} containing the CAR \eqn{(\rho_k)_k} values.}

\item{shape}{shape value of the inverse gamma assumed a priori.}

\item{scale}{scale value of the inverse gamma assumed a priori.}
}
\value{
a value generated from the conditional posterior distribution of \eqn{\tau}.
}
\description{
This function update the CAR variance parameter drawing from its conditional posterior distributions, which is Inverge Gamma.
}
