% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tau_metropolis.R
\name{update_tau_metropolis}
\alias{update_tau_metropolis}
\title{Posterior update of the CAR variance parameters via Metropolis-Hastings}
\usage{
update_tau_metropolis(tau_current, Psi, D, W, Rho, shape, scale, eps = 1)
}
\arguments{
\item{tau_current}{the current parameter value.}

\item{Psi}{a matrix of dimension \verb{n x (K-1)}, where \code{n} is the number of spatial areas in the analysed map, and \code{K} is the number of clusters assumed.}

\item{D}{the diagonal matrix containing the number of neighbours.}

\item{W}{the proximity matrix.}

\item{Rho}{a vector of length \code{K-1} containing the CAR \eqn{(\rho_k)_k} values.}

\item{shape}{shape value of the inverse gamma assumed a priori.}

\item{scale}{scale value of the inverse gamma assumed a priori.}

\item{eps}{a tuning parameter (default \code{1}). Proposed values are generated from \code{rlnorm} with \code{meanlog = log(tau_current)} and \code{sdlog = eps}.}
}
\value{
A list of two elements:
\itemize{
\item \code{value} contains the value simulated by the algorithm;
\item \code{accepted} is a value that equals 1 if the returned value is different from \code{tau_current}, and 0 otherwise.
}
}
\description{
This function update the CAR variance parameter drawing proposing values from a log-Normal distribution
}
