% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_clusters.R
\name{update_Z}
\alias{update_Z}
\title{Update the clustering labels
This function updates the clustering labels of the \code{n} areas of the map}
\usage{
update_Z(y, mu, Sigma, Psi)
}
\arguments{
\item{Psi}{}
}
\description{
Update the clustering labels
This function updates the clustering labels of the \code{n} areas of the map
}
