% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{west_states_data}
\alias{west_states_data}
\title{Data on mortality by U.S. counties (\code{SpatialPolygonDataFrame} object)}
\format{
\subsection{\code{west_states_data}}{

The file contains a \code{SpatialPolygonDataFrame} object with the map and the
data of the mortality by U.S. counties from 2016 to 2019 for three causes
of death (circulatory diseases, respiratory diseases and tumors).
}
}
\source{
{WONDER Online Database (https://wonder.cdc.gov/mcd-icd10.html).}
}
\usage{
west_states_data
}
\description{
Data on mortality by U.S. counties from 2016 to 2019 for the West Coast.
Data are available in the WONDER Online Database at
https://wonder.cdc.gov/mcd-icd10.html.
}
\examples{
library(sp)
data(west_states_data)
plot(west_states_data)

}
\keyword{datasets}
