% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate.simulations}
\alias{generate.simulations}
\title{Simulate maps using DAGAR}
\usage{
generate.simulations(
  spatial.map,
  K,
  d = NULL,
  Sigma = NULL,
  range.mu,
  rho,
  prob.null.centroid = 0.5,
  scale.factor.variance = 0.05,
  correct.mean.prob = F,
  plot.map = F
)
}
\arguments{
\item{spatial.map}{A \code{SpatialPolygonsDataFrame} object with the map you want
to simulate from.}

\item{K}{Number of clusters.}

\item{d}{Number of the events that are inducing the clusters.}

\item{Sigma}{Correlation matrix across the events that are inducing the
clusters.}

\item{range.mu}{Range of the uniform distribution used in data generation.}

\item{rho}{\code{rho} parameter of the DAGAR.}

\item{prob.null.centroid}{Probability of the uniform distribution used in
data generation.}

\item{scale.factor.variance}{Scaling factor for variance-covariance matrix (
default 0.05).}

\item{correct.mean.prob}{Use the stick-breaking formulation of the
multinomial distribution while simulating (default \code{FALSE}).}

\item{plot.map}{Plot the map during the simulation (default \code{FALSE}).}
}
\value{
The function returns the map and the simulated data.
}
\description{
Simulate maps using DAGAR.
}
\examples{
west.states <- data(west_states_data)

sim1 <- generate.simulations(spatial.map = west.states,
K = 3,
d = 10,
Sigma = NULL,
range.mu = 0.5,
rho = c(0.8, 0.9),
prob.null.centroid = 0.5,
scale.factor.variance = 0.05,
correct.mean.prob = F,
plot.map = F)

Corr.matrix <- matrix(c(0.07, -0.01175893, 0.00124239,
-0.01175893, 0.07, 0.00653324,
0.00124239, 0.00653324, 0.07),
byrow = T, ncol = 3, nrow = 3)

sim2 <- generate.simulations(spatial.map = west.states,
K = 3,
d = NULL,
Sigma = Corr.matrix,
range.mu = 1, rho = c(0.01, 0.9),
prob.null.centroid = 0.5,
scale.factor.variance = 0.05,
correct.mean.prob = F,
plot.map = F)

}
