% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationModel.R
\name{generate.simulations}
\alias{generate.simulations}
\title{Simulate maps}
\usage{
generate.simulations(
  map,
  K,
  d = NULL,
  Sigma = NULL,
  range.mu,
  rho,
  simulation.model = "dagar",
  tau = NULL,
  allocation = "deter",
  prob.null.centroid = 0.5,
  scale.factor.variance = 0.005,
  correct.mean.prob = F,
  plot.map = F
)
}
\arguments{
\item{plot.map}{}
}
\description{
Simulate maps
}
