% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{information.criteria}
\alias{information.criteria}
\title{Title}
\usage{
information.criteria(values, loglikelihood.values = NULL)
}
\arguments{
\item{values}{an object of class \code{perla}}

\item{loglikelihood.values}{}
}
\description{
Title
}
