% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_perla.R
\name{merge.perla}
\alias{merge.perla}
\title{Merge multiple perla objects into a unique chain}
\usage{
merge.perla(values)
}
\arguments{
\item{values}{a list of \code{perla} objects. They must be different parallel runs of the same model.}
}
\value{
An object of class perla. It contains an object called \code{chains}, which states from which chain the elements come from.
}
\description{
This function is designed to combine multiple Markov chains obtained running Perla from different starting points.
}
