% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perla.R
\name{perla}
\alias{perla}
\title{PEnalized Regression with Local Aggregation (PERLA)}
\usage{
perla(
  x,
  K,
  R = 10^4,
  prior.rho = "const",
  p.spike = 0.5,
  rho.value = 0.99,
  mu0 = NULL,
  Sigma0 = NULL
)
}
\arguments{
\item{x}{an \code{n} x \code{d} matrix, where \code{n} is the number of spatial areas and \code{d} is the number of diseases}

\item{K}{the number of clusters}

\item{R}{the number of MCMC iterations (default \code{10^4})}

\item{prior.rho}{set the type of prior on the rho parameter of the CAR. If \code{const}, it assumes a fixed value, that is \code{rho.value}.
If \code{disc}, it assumes that rho can be equal to \code{0} (with probability \code{p.spike}) or \code{rho.value}.
If \code{cont}, it assumes that rho is generated from a mixture of a \code{dbeta(x,2,18)} (with probability \code{p.spike}) or a \code{dbeta(x,18,2)}.}

\item{Sigma0}{}
}
\description{
This function estimates the PERLA model using the Markov Chain Monte Carlo algorithm.
}
