% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive.R
\name{posterior_predictive_sampling}
\alias{posterior_predictive_sampling}
\title{Generate values from the posterior predictive distribution
This function generate a new dataset with the same dimension of the analyzed one from the posterior predictive distribution}
\usage{
posterior_predictive_sampling(Z, Mu, Sigma)
}
\arguments{
\item{Sigma}{}
}
\description{
Generate values from the posterior predictive distribution
This function generate a new dataset with the same dimension of the analyzed one from the posterior predictive distribution
}
