% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{recover.loglikelihood}
\alias{recover.loglikelihood}
\title{Log-likelihood computation}
\usage{
recover.loglikelihood(values, burnin = NULL)
}
\arguments{
\item{values}{an object of class \code{perla}}

\item{burnin}{a vector of indexes denoting the MCMC draws to be discarded (default \code{NULL} means every draw is kept)}
}
\value{
an object of class \code{perla} that contains also \code{loglik}, the vector of log-likelihood values for every MCMC itereation.
}
\description{
This function returns the log-likelihood value at each iteration of the MCMC.
}
